Crate blockz_derive

Blockz derive.

Modules

common
Common functions.

configuration	configuration
#[derive(Configuration)].

errors
Errors.

factory
Factories that build code.

paths
Paths to libraries used by the expansion of blockz proc-macros.

singleton	singleton
Singleton macro.

Attribute Macros

singleton_fn	singleton
Modify a method on a type that implements the Singleton trait.

Derive Macros

Configuration	configuration
Derive the Configuration trait.

Singleton	singleton
Derive the Singleton trait.