Enum blockz_derive::singleton::lock::SingletonLock

pub(super) enum SingletonLock {
    Mutex,
    RwLock,
}

This is supported on crate feature singleton only.

The lock behing a singleton.

Variants

Mutex

tokio::sync::Mutex

RwLock

tokio::sync::RwLock

Implementations

impl SingletonLock

pub fn to_type(&self, inner: &Ident) -> TokenStream

Create a new lock type with the inner type.

pub fn to_new_lock_expr(&self, inner: &Ident) -> TokenStream

Create a lock initialization expression that sets the inner value to the ident.

pub fn to_guard(&self, lock: &Ident) -> TokenStream

Create a new lock guard from a lock identifier.

pub fn to_mut_guard(&self, lock: &Ident) -> TokenStream

Create a new mutable lock guard from a lock identifier.