Enum blockz_derive::singleton::singleton_fns::SingletonFnArgs

pub(super) enum SingletonFnArgs<'f> {
    Single {
        arg_ident: &'f Ident,
        arg_type: &'f Type,
    },
    Multiple {
        arg_idents: Vec<&'f Ident>,
        tuple_ident: Ident,
        tuple_type: TypeTuple,
    },
}

This is supported on crate feature singleton only.

Arguments for a SingletonFn.

Variants

Single

Single argument.

This is passed as-is to the singleton function.

Fields of Single

arg_ident: &'f Ident

The identifier of the function argument.

arg_type: &'f Type

The type of the function argument.

Multiple

Multiple arguments.

These need to be converted to a single argument as a tuple.

Fields of Multiple

arg_idents: Vec<&'f Ident>

The idents of the original function arguments.

tuple_ident: Ident

The ident of the tuple function argument.

tuple_type: TypeTuple

The type of the tuple function argument.

This is just a combination of the original types of the tuple function argument.

Implementations

impl<'f> SingletonFnArgs<'f>

pub fn build_impl_fn_call_arg(&self) -> TokenStream

Build the argument used for a singleton call by the facade fn.

This will either just return the ident of the arg, if there is only one, or it will return the code needed to build a tuple arg.

The call arg will be used for: Singleton::use_singleton_with_arg or Singleton::use_mut_singleton_with_arg

pub fn build_impl_fn_sig_arg(&self) -> Result<FnArg>

Build the fn input that will be accepted by the impl fn.

This builds the arg: (i64, u64, ...) in the function signature.

pub fn build_impl_fn_replacement_legend(
&self
) -> Option<HashMap<String, TokenStream>>